move search_cutoff default logic
