+++ src/loki/isochrone_action.cc
@@ -33,7 +33,7 @@ void loki_worker_t::init_isochrones(Api& request) {
   auto& options = *request.mutable_options();
 
   // strip off unused information
-  parse_locations(options.mutable_locations());
+  parse_locations(options.mutable_locations(), request);
   if (options.locations_size() < 1) {
     throw valhalla_exception_t{120};
   };
@@ -75,7 +75,7 @@ void loki_worker_t::isochrones(Api& request) {
 
   try {
     // correlate the various locations to the underlying graph
-    auto locations = PathLocation::fromPBF(request, options.locations());
+    auto locations = PathLocation::fromPBF(options.locations());
     const auto projections = loki::Search(locations, *reader, costing);
     for (size_t i = 0; i < locations.size(); ++i) {
       const auto& projection = projections.at(locations[i]);
