+++ src/loki/locate_action.cc
@@ -9,7 +9,7 @@ namespace valhalla {
 namespace loki {
 
 void loki_worker_t::init_locate(Api& request) {
-  parse_locations(request.mutable_options()->mutable_locations());
+  parse_locations(request.mutable_options()->mutable_locations(), request);
   if (request.options().locations_size() < 1)
     throw valhalla_exception_t{120};
 
@@ -22,7 +22,7 @@ std::string loki_worker_t::locate(Api& request) {
 
   // correlate the various locations to the underlying graph
   init_locate(request);
-  auto locations = PathLocation::fromPBF(request, request.options().locations());
+  auto locations = PathLocation::fromPBF(request.options().locations());
   auto projections = loki::Search(locations, *reader, costing);
   return tyr::serializeLocate(request, locations, projections, *reader);
 }
